pyproject.toml
[tool.poetry]
name = "gcsv"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
google-api-python-client = "^1.7"
google-auth-httplib2 = "^0.0.3"
google-auth-oauthlib = "^0.2.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
